[CMP_DAC] Regenerated demo for cmp_dac__cmp_dac_sine_wave__pic32ak_gc_curiosity.X
MH3-110677

+++ apps/cmp_dac/cmp_dac_sine_wave/src/config/default/peripheral/cmp_dac/plib_cmp1.h
@@ -17,7 +17,7 @@
 *******************************************************************************/
  
 /*******************************************************************************
-* Copyright (C) 2024 Microchip Technology Inc. and its subsidiaries.
+* Copyright (C) 2025 Microchip Technology Inc. and its subsidiaries.
 *
 * Subject to your compliance with these terms, you may use Microchip software
 * and any derivatives exclusively with Microchip products. It is your
@@ -137,7 +137,7 @@ inline static void CMP1_DACDisable(void)
  */
 inline static void CMP1_DACDataWrite(size_t value)
 {
-    DAC1DATbits.DACDAT = value;
+    DAC1DATbits.DACDAT = (uint16_t)value;
 }
